News can be converted to HTML and FB2 formats
Two new arguments to-html and to-fb2 are added to convert news to HTML and FIctionBook2 formats respectively.
Value of these arguments is a new file path to store convertion result. File with such name should not exist. Otherwise, an error occurs.
If image is skipped in news, default one is used. As fb2 supports on jpeg and png formats, default image is used instead of all other formats.
There are also several changes\fixes:
 - setup.py contains required additonal packages as requirements.txt
 - cached news\images are stored in "home"/.rssreader folder